Refactor some code of Enemy and Player
